requirements.txt
python-dotenv
fastapi[all]
uvicorn
websockets
loguru
numpy
requests
pipecat-ai[google,silero,webrtc]
